+++ src/config.ts
@@ -1,6 +1,9 @@
 enum Command {
     Search = "projecteuler.searchProblem",
     Show = "projecteuler.showProblem",
+    Refresh = "projecteuler.refreshTree",
+    Clear = "projecteuler.clearTreeData",
+    Load = "projecteuler.loadProblemData",
 }
 
 export { Command };