+++ src/service.ts
@@ -1,5 +1,7 @@
 import * as vscode from "vscode";
+import * as cheerio from "cheerio";
 import Papa from "papaparse";
+import { Command } from "./config";
 
 export interface ProblemData {
     "ID": number,
@@ -10,6 +12,7 @@ export interface ProblemData {
 }
 
 export class ProblemDataService {
+    public static basePath: vscode.Uri = vscode.Uri.parse("https://projecteuler.net");
     private static problemInfoKey: string = "euler.problem-info";
     private context: vscode.ExtensionContext;
 
@@ -17,26 +20,80 @@ export class ProblemDataService {
         this.context = context;
     }
 
-    public getProblemInfo(): { [key: number]: ProblemData } {
-        return this.context.globalState.get(ProblemDataService.problemInfoKey) as { [key: number]: ProblemData };
+    public clearProblemInfo(): void {
+        this.context.globalState.update(ProblemDataService.problemInfoKey, undefined);
+        vscode.commands.executeCommand(Command.Refresh);
     }
 
-    public async getProblemData(id: string | number): Promise<string> {
-        let response = await this._fetchProblem(id);
-        return new Promise<string>((resolve, reject) => {
-            if (response.status !== 200) {
-                reject();
+    public getProblemInfo(): { [key: number]: ProblemData } {
+        const data = this.context.globalState.get(ProblemDataService.problemInfoKey);
+        if (data) {
+            return data as { [key: number]: ProblemData };
+        } else {
+            return {};
+        }
+    }
+
+    private async _download(source: vscode.Uri, path: string): Promise<vscode.Uri> {
+        let target = vscode.Uri.joinPath(this.context.globalStorageUri, path);
+        try {
+            await vscode.workspace.fs.stat(target);
+        } catch {
+            let response = await fetch(source.toString());
+            let content = await response.bytes();
+            vscode.workspace.fs.writeFile(target, content);
+        }
+        return target;
+    }
+
+    private async _downloadContent(text: string, panel: vscode.WebviewPanel): Promise<string> {
+        const $ = cheerio.load(`<div class="problem_content">${text}</div>`);
+        {
+            let images = $("img");
+            for (let i = 0; i < images.length; i += 1) {
+                if (images[i].attribs.src && images[i].attribs.src.startsWith("resources")) {
+                    let path = vscode.Uri.parse(images[i].attribs.src).path;
+                    let uri = vscode.Uri.joinPath(ProblemDataService.basePath, path);
+                    let target = await this._download(uri, path);
+                    let srcUri = panel.webview.asWebviewUri(target);
+                    images[i].attribs.src = `${srcUri}`;
+                }
             }
-            response.text()
-                .then((text) => {
-                    if (text !== "Data for that problem cannot be found") {
-                        resolve(text);
-                    } else {
-                        reject();
+        }
+
+        {
+            let links = $("a");
+            for (let i = 0; i < links.length; i += 1) {
+                let href = links[i].attribs.href;
+                if (href) {
+                    if (href.startsWith("about")) {
+                        links[i].attribs.href = `${vscode.Uri.joinPath(ProblemDataService.basePath, href)}`;
+                    } else if (href.startsWith("resources")) {
+                        let path = vscode.Uri.parse(links[i].attribs.href).path;
+                        let uri = vscode.Uri.joinPath(ProblemDataService.basePath, path);
+                        await this._download(uri, path);
+                        links[i].attribs.class = "linked-resource";
                     }
-                })
-                .catch((_) => reject());
-        });
+                }
+            }
+        }
+
+        return $.html();
+    }
+
+    public async getProblemData(id: string | number, panel: vscode.WebviewPanel): Promise<string> {
+        let target = vscode.Uri.joinPath(this.context.globalStorageUri, "problems", `${id}.html`);
+        try {
+            await vscode.workspace.fs.stat(target);
+            const problemData = await vscode.workspace.fs.readFile(target);
+            return `${problemData}`;
+        } catch {
+            const response = await this._fetchProblem(id);
+            const text = await response.text();
+            const problemHtml = await this._downloadContent(text, panel);
+            vscode.workspace.fs.writeFile(target, Buffer.from(problemHtml));
+            return problemHtml;
+        }
     }
 
     public getTitle(id: number | string): string {
@@ -69,5 +126,6 @@ export class ProblemDataService {
         }
 
         this.context.globalState.update(ProblemDataService.problemInfoKey, problemInfo);
+        vscode.commands.executeCommand(Command.Refresh);
     }
 }